After handler brought up to speed.
As discussed in issue #151, I modified the implementation to be simpler.
I checked how to propagate err, being done with a scope trick. From our
deferred function we cannot return a new value, so we override the "default"
return value (err), declared on functions' declarations.

Check this StackOverflow for more info: http://stackoverflow.com/a/19934989