app: Fix trailing space for Author.String()
This code initially landed with lots of space:

  '{name}  <{email}> '

or:

  '{name} '

in 21e18a14 (app, help: add support for multiple authors, 2015-01-31).
The doubled space between the name and email was removed in f45dd049
(app, help: add backwards compatibility for Authors, 2015-02-21), but
a trailing space remained in both the email and email-less cases.
This commit removes that trailing space.