Make exit code example more clear
The purpose of this example is to show that you can exit with an error
code if a flag is unspecified, but with the code as it is, the only way
to cause a non-zero exit is by adding the flag `--ginger-crouton=false`,
which is not explained in the example.

In this new version of the example, running the command with no flag
will exit with an error, and running it with the flag will exit
normally.